Change to protocol-less URI
The protocol check is not needed when the only part of the URI that changes is the protocol.

Can simply use _//ajax.googleapis.com/ajax/libs/jquery/1.5.1/jquery.min.js_ as the src attribute and it will work in both HTTP and HTTPS cases.

Reference: [section 4.2 of RFC 3986](http://tools.ietf.org/html/rfc3986#section-4.2)